moves dht to its own package
