+++ network/p2p/dht/dht.go
@@ -1,4 +1,4 @@
-package p2p
+package dht
 
 import (
 	"context"
