+++ network/test/blob_service_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/ipfs/go-datastore"
 	"github.com/ipfs/go-datastore/sync"
 	blockstore "github.com/ipfs/go-ipfs-blockstore"
+	"github.com/onflow/flow-go/network/p2p/dht"
 	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/atomic"
@@ -23,7 +24,6 @@ import (
 	"github.com/onflow/flow-go/network"
 	"github.com/onflow/flow-go/network/channels"
 	"github.com/onflow/flow-go/network/mocknetwork"
-	"github.com/onflow/flow-go/network/p2p"
 )
 
 // conditionalTopology is a topology that behaves like the underlying topology when the condition is true,
@@ -86,7 +86,7 @@ func (suite *BlobServiceTestSuite) SetupTest() {
 		logger,
 		unittest.NetworkCodec(),
 		mocknetwork.NewViolationsConsumer(suite.T()),
-		WithDHT("blob_service_test", p2p.AsServer()),
+		WithDHT("blob_service_test", dht.AsServer()),
 		WithPeerUpdateInterval(time.Second),
 	)
 	suite.networks = networks
